Create CalendarService and inject it
