+++ LTUAssistantPlus/skills/add_calendar_event_skill.py
@@ -1,9 +1,8 @@
 #!/usr/bin/python3
 
-import calendardb
-
 from nlp.universal_dependencies import ParsedUniversalDependencies
 from services.assistant_services import AssistantServices
+from services.calendar_service import CalendarEvent
 from .skill import SkillInput, Skill
 
 class AddCalendarEventSkill(Skill):
@@ -26,8 +25,8 @@ class AddCalendarEventSkill(Skill):
             event_sentence = services.user_interaction_service.ask_question('Okay, what is the event called?', skill_input.verbose)
             day_sentence = services.user_interaction_service.ask_question('What day will this be on?', skill_input.verbose)
             time_sentence = services.user_interaction_service.ask_question('What time will this start at?', skill_input.verbose)
-            cal_event = calendardb.CalendarEvent(event_sentence, day_sentence, time_sentence, '')
-            calendardb.add_event(cal_event)
+            cal_event = CalendarEvent(event_sentence, day_sentence, time_sentence, '')
+            services.calendar_service.add_event(cal_event)
             feedback_sentence = 'Alright, I\'m putting down ' + str(cal_event) + '.'
             services.user_interaction_service.speak(feedback_sentence, skill_input.verbose)
         else:
